Use a matchspec trick to only send args if capturing is on
Use 'arity' trace flag to avoid sending arguments of function calls in
all trace messages (which can be potentially big terms). If argument and
return value capturing is on use a special match spec for that
particular function. This match spec uses the action body 'message' to
return the arguments as a separate field in the trace message.

This way sending args can be turned on selectively per mfa.
Unfortunately this does not work for return values which are always
included in the 'return_from' trace messages.