(#19) Auto-complete improvements [ENTER]
Enter either submits selected suggestion if there is any or content of
textbox.

It is checked if textbox content is the prefix of the selected fun to
make sure we don't override content which is different from the
selection. One case when this can happen is when a match-spec fun is
under editing, and there is exactly one suggestion which is
auto-selected. In this case we don't want to submit mfa with arity.