(#22) "api/mon_get_all" endpoint now also returns formatted MFA
The extended mfa array is transparent to the functions which fetch mod,
fun and arity from the first three elements.
  [mod, fun, arity] -> [mod, fun, arity, fmtMFA]

This new field (which is the original query string) is used as graph
title and basis of chart_id. The javascript implementation of formatMFA
is not needed any more.

The following fields contain the new data structure
- GraphPanel.state.funs
- Graph.props.mfa
- CallTracer.props.mfa