Bugfix calls longer than 1 second were silently ignored
Calls that took more than 1 second were although visible when data
capturing is enabled they were not recorded by
hdr_histogram. hdr_histogram is started with a maximum value which was
set to 1e6 (microseconds) and when trying to store larger values the
`record/2` call returned {error, value_out_of_range} which was ignored.

Now increase the max value to 30 seconds, and in case a call takes even
longer than this, store the max value and emit an error log.