(#22) More error handling running elixir tests
It might be possible that a certain version of elixir is available but
there is some problem invoking it. (For example on travis there is a
default elixir 1.0.2 which is so old it crashes on startup with recent
erlang versions (at least from R16B03-1))
Skip elixir tests in this case as well.