Fix edge case in match-spec rewrite
We modify the match-spec so that if we don't capture arguments no unused
term is included in the trace message. However some tricky
folks (eg. me) include logic within the `message()` action-function to
conditionally mute tracing. This was just replaced all together.

Now this is fixed by not only checking if the value withing `message()`
is literarily `false` but also whether it evaluates to `false` (at
run-time)