Improve error handling of capture functions
Handle cases when MFA is not monitored or not captured.

As an exception in case of `get_captured_data` return default
values (-1) if MFA is monitored but not captured as this is polled by
the frontend and used as a kind of capture status. This could be
improved in the future.