Move plugins to profiles
Move hex plugin to its own, new publish profile. It is only used for
creating hex pacakges but not for fetching them. It uses maps and cannot
be compiled on R16B. This is not an issue just creates noise in the
output.

While at it move coveralls plugin to the test profile as it is only
used in test commands.