Don't keep the same entry in recent queries multiple times
Before this it was by-design that every successful query was present in
the list of recent queries (just like in bash history). But in this
particular use-case I think it is more user-friendly to keep only the
latest call of the exact same queries - this way the list will be
shorter and easier to scroll.