Support typed records
We use `compile:forms` to convert a fun to a match-spec using the
`ms_transorm` parse-transform. To also recognise records the definitions
are fetched from the source module. However the definitions might
contain typed fields and local user type definitions are not fetched.

The `P` option of the compiler also runs the linter after module
tranformation which complains about an undefined type. Instead use the
`dpp` compiler option which solely runs the tranformation.