Format exception in OTP 21
In OTP 21 `format_exception/7` was moved from `lib` module to `erl_error`

We need to enable edoc option `preprocess` to generate edoc on OTP 20 or
older. By default edoc tries to parse the file without preprocessing,
ignores the directive `-ifdef(OTP_RELEASE).` and fails on `-if(...).` in
the next line as the if directive was only introduced in OTP 21.

From the documentation of edoc:
> Normally, preprocessing is not necessary for EDoc to work, but if a
> file contains too exotic definitions or uses of macros, it will not be
> possible to read it without preprocessing.