Improve pretty-printing and line-breaking of erlang terms
The default formatter does not insert spaces except in records and
maps.

There was an issue that in captured arguments and return values
the terms wouldn't fit in their column on the GUI if it was narrower
than 80 characters, because the browser couldn't break the long strings
without spaces.

With this commit we insert spaces everywhere (where it is common in code
style), so the output will be more spacy and better breakable.