Support typed records - second try
We use `compile:forms` to convert a fun to a match-spec using the
`ms_transorm` parse-transform. To also recognise records the
record definitions are fetched from the source module and
included in the compiled forms. However the definitions might
contain typed fields referencing local user type definitions
(which are not included) and the linter would emmit
an error on these unknown local types.

Previous attemp to address this (in commit adca1349) changed the
`P` compiler option to `dpp` which does not run the linter.
However we still want to catch and pretty pring other errors
(like unbound variable usage).

So this next attempt instead removes the type annotation from
record defs when they are stored in ETS, and restores calling the
compiler with the 'P' option.

closes #26