Add poor man's notifications
Store pid/regname of event handler process in application env.

A message in the form `{snapshost, NowTime}` is sent to the event
handler when a snaphost is created.

The testcases subscribe themselves as event handlers making them
synchornous and bit less prone to timing issues.

So far this is only used for tests but could be the start of a
websocket-based notification system.