Restructure ifdefs
I was hoping to jump to OTP 19+ support soon (with tracer NIFs etc), but
it looks like XProf will need to support OTP 18 for a while. So let's
use all the new functions (like rand module, math:log2, erl_anno etc) if
they are available, and create a single `before_OTP_18` macro for the
older stuff.